+++ packages/admin-ui/ui/src/domain/inventory/header.tsx
@@ -1,8 +1,12 @@
 import TableViewHeader from "../../components/organisms/custom-table-header"
 import { useNavigate } from "react-router-dom"
+import {
+  FeatureFlag,
+  useFeatureFlag,
+} from "../../providers/feature-flag-provider"
 
 type P = {
-  activeView: "inventory" | "locations"
+  activeView: "inventory" | "locations" | "reservations"
 }
 
 /*
@@ -11,16 +15,27 @@ type P = {
 
 function InventoryPageTableHeader(props: P) {
   const navigate = useNavigate()
+  const { isFeatureEnabled } = useFeatureFlag()
+  const isReservationsEnabled =
+    isFeatureEnabled(FeatureFlag.RESERVATIONS) &&
+    isFeatureEnabled(FeatureFlag.INVENTORY)
+
+  const views = ["inventory", "locations"]
+
+  if (isReservationsEnabled) {
+    views.push("reservations")
+  }
+
   return (
     <TableViewHeader
       setActiveView={(v) => {
         if (v === "inventory") {
           navigate(`/a/inventory`)
         } else {
-          navigate(`/a/inventory/locations`)
+          navigate(`/a/inventory/${v}`)
         }
       }}
-      views={["inventory", "locations"]}
+      views={views}
       activeView={props.activeView}
     />
   )