+++ packages/modules-sdk/src/loaders/__tests__/module-loader.ts
@@ -1,22 +1,12 @@
 import {
-  ModuleResolution,
   MODULE_RESOURCE_TYPE,
   MODULE_SCOPE,
+  ModuleResolution,
 } from "@medusajs/types"
-import { AwilixContainer, ClassOrFunctionReturning, Resolver } from "awilix"
 import { createMedusaContainer } from "medusa-core-utils"
 import { EOL } from "os"
-import { moduleLoader } from "../module-loader"
 import { trackInstallation } from "../__mocks__/medusa-telemetry"
-
-function asArray(
-  resolvers: (ClassOrFunctionReturning<unknown> | Resolver<unknown>)[]
-): { resolve: (container: AwilixContainer) => unknown[] } {
-  return {
-    resolve: (container: AwilixContainer): unknown[] =>
-      resolvers.map((resolver) => container.build(resolver)),
-  }
-}
+import { moduleLoader } from "../module-loader"
 
 const logger = {
   warn: jest.fn(),